Removes unused API routes.
In order to perform SSG on dynamic and static routes, we need to create fetch data directly from third party APIs (Sanity & Salesforce).
The issue with using Next.js API routes is that they do not exist until after build time, which means we cannot use them to fetch data for SSG.

This commit also fixes various link typos and unused imports from the codebase.
