+++ services/internshipPageService.tsx
@@ -1,19 +1,25 @@
-import { API_ENDPOINTS } from '@/constants/api'
-import { api, RequestType } from '@/services/api';
 import { InternshipActionDataProps } from '@/components/Internship/internshipblogsection/internshipblogsection';
 import { IntershipBenefitDataProps } from '@/components/Internship/Interashipdetails/Interashipdetails';
 import { IntershipOfferDataProps } from '@/components/Internship/receiveoffcourses/receiveoffcourses';
+import { client } from '@/sanity/lib/client';
+import { urlForImage } from '@/sanity/lib/image';
 
 const internshipPageService = {
     fetchActionItem: async (): Promise<InternshipActionDataProps[]> => {
         try {
-            const response = await api({
-                endpoint: API_ENDPOINTS.fetchActionItem,
-                type: RequestType.GET,
+            const actionItems = await client.fetch<InternshipActionDataProps[]>(`*[_type == "internship_action"]`);
+
+            actionItems.forEach((item: InternshipActionDataProps) => {
+                if (item.action_image?.asset?._ref) {
+                    const imageUrl = urlForImage(item.action_image.asset);
+                    if (item.action_image.asset) {
+                        item.action_image.asset.image_url = imageUrl;
+                    }
+                }
             });
 
-            if (response?.status === 200) {
-                return response.data as InternshipActionDataProps[];
+            if (actionItems) {
+                return actionItems as InternshipActionDataProps[];
             } else {
                 throw new Error('Failed to fetch Action Items');
             }
@@ -24,13 +30,14 @@ const internshipPageService = {
     },
     fetchInternshipBenefits: async (): Promise<IntershipBenefitDataProps> => {
         try {
-            const response = await api({
-                endpoint: API_ENDPOINTS.fetchInternshipBenefits,
-                type: RequestType.GET,
-            });
+            const benefitItems = await client.fetch(`*[_type == "internship_benefits"][0]`);
 
-            if (response?.status === 200) {
-                return response.data as IntershipBenefitDataProps;
+            if (benefitItems.logo?.asset?._ref) {
+                benefitItems.logo.asset.image_url = urlForImage(benefitItems.logo.asset); // Add the image URL to the response
+            }
+
+            if (benefitItems) {
+                return benefitItems as IntershipBenefitDataProps;
             } else {
                 throw new Error('Failed to fetch Internship Benefits');
             }
@@ -41,13 +48,10 @@ const internshipPageService = {
     },
     fetchInternshipOffer: async (): Promise<IntershipOfferDataProps> => {
         try {
-            const response = await api({
-                endpoint: API_ENDPOINTS.fetchInternshipOffer,
-                type: RequestType.GET,
-            });
+            const benefitItems = await client.fetch(`*[_type == "internship_offer"][0]`);
 
-            if (response?.status === 200) {
-                return response.data as IntershipOfferDataProps;
+            if (benefitItems) {
+                return benefitItems as IntershipOfferDataProps;
             } else {
                 throw new Error('Failed to fetch Internship Benefits');
             }
